refactor(music): migrate to musicFormat for formatting and timing helpers, remove musicPlayerV2
